Move capture output in its own class for performance
